requirements.txt
tensorflow-gpu==1.7
pandas
xlrd
matplotlib
